Update SAMPLE.sol
This smart contract provides a daily return of 0.25% on deposits and has a referral bonus of 3% on the initial deposit sent to referrers. There is also a 3% development/marketing fee on deposits and withdrawals. 25% of deposits are sent to a trading (salsa) wallet, and users can compound their rewards twice, with a mandatory waiting period of 24 hours between compounds. There is a 30 hour withdrawal window after the last compound, after which rewards will stop accumulating until a compound or withdrawal is made.
Additionally, users can deposit funds into the salsa pool, which are then invested into other investments and the user's share of rewards from the investment is credited to the user's rewards balance in the main balance.
A 1% fee for the management and for gas fees, are sent to the owner's wallet upon deposit and each time the owner deposits investment rewards.
Finally, the contract has a lottery feature, where a user can buy a lottery ticket with a click and the price is set to 10 BUSD. Once 10 tickets are purchased, 50 BUSD is sent to a random winner and 50 BUSD is kept in the contract. The last lottery winner is stored and replaced each time the lottery finishes.